refactor(mearie)!: remove mearie/config export path
BREAKING CHANGE: The mearie/config export path has been removed. Import defineConfig directly from 'mearie' instead of 'mearie/config'.
