+++ .changeset/proud-lions-dance.md
@@ -1,0 +1,5 @@
+---
+'mearie': major
+---
+
+Remove `mearie/config` export path in favor of importing `defineConfig` directly from `mearie`
