+++ examples/basic/mearie.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from 'mearie/config';
+import { defineConfig } from 'mearie';
 
 export default defineConfig({
   scalars: {
