--- packages/mearie/src/config.ts
@@ -1,1 +1,0 @@
-export { defineConfig, type MearieConfig, type ResolvedMearieConfig } from '@mearie/config';
